pyproject.toml
[project]
name = "alfakids-backend"
version = "0.1.0"
description = "Backend for AlfaKids"
authors = [
    { name = "Felipe Martins", email = "felipemedeirosinfo@gmail.com.com" }
#    { name = "Anderson", email = "felipemedeirosinfo@gmail.com.com" }
]
dependencies = [
    "Flask~=2.3.3",
    "alembic~=1.14.0",
    "SQLAlchemy~=2.0.36",
    "Flask-SQLAlchemy~=3.0.0",
    "Flask-Migrate~=4.0.4",
    "psycopg2-binary~=2.9.8",
    "Flask-Cors~=4.0.0",
    "Flask-JWT-Extended~=4.4.4",
    "requests~=2.26.0",
    "bcrypt~=4.2.1",
    "python-dotenv~=1.0.1",
    "Flask-Mail~=0.10.0"
]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
